feat: provide clang-cl support

+++ src/sentry_utils.h
@@ -154,7 +154,7 @@ sentry__monotonic_time(void)
 
     LARGE_INTEGER qpc_counter;
     QueryPerformanceCounter(&qpc_counter);
-    return qpc_counter.QuadPart * 1000 / qpc_frequency.QuadPart;
+    return (uint64_t)qpc_counter.QuadPart * 1000 / (uint64_t)qpc_frequency.QuadPart;
 #elif defined(SENTRY_PLATFORM_DARWIN)
 
 // try `clock_gettime` first if available,
